+++ src/llmreplay/lineage/__init__.py
@@ -1,0 +1,16 @@
+"""Lineage / tweak / sticky / templates package."""
+
+from llmreplay.lineage.fork import fork_cassette, load_lineage
+from llmreplay.lineage.sticky import maybe_sticky_write, sticky_writeback_allowed
+from llmreplay.lineage.templates import apply_materializer, list_materializers
+from llmreplay.lineage.tweak import tweak_transaction
+
+__all__ = [
+    "apply_materializer",
+    "fork_cassette",
+    "list_materializers",
+    "load_lineage",
+    "maybe_sticky_write",
+    "sticky_writeback_allowed",
+    "tweak_transaction",
+]